MdeModulePkg: Allocate Io Queue buffer after setting the number of queues with the controller.
This commit refactors NvmeControllerInit() in NvmExpressHci.c.
First it was modified to contain the logic for allocating the I/O queue
buffer after the number of queues has been set with the controller.
This change ensures that the I/O queue buffer is allocated with the
correct.

Because NvmeControllerInit() was used for a controller reset,
a new function, NvmeControllerReset(), was created to only
apply the reset logic. For example, on reset the Admin Queue
Attributes should stay the same. The queue buffer allocations
can also remain the same. In the new reset function, more
checks are added to ensure the AQA and Number of Queues haven't changed.

Signed-off-by: Vivian Nowka-Keane <vnowkakeane@linux.microsoft.com>